accounts/api/urls.py
from django.urls import include, path
from rest_framework_simplejwt.views import (
    TokenObtainPairView,
    TokenRefreshView,
    TokenVerifyView,
)

from .views import UserDetailAPIView

app_name = "accounts"

jwt_urlpatterns = [
    path("token/", TokenObtainPairView.as_view(), name="token_obtain_pair"),
    path("token/refresh/", TokenRefreshView.as_view(), name="token_refresh"),
    path("token/verify/", TokenVerifyView.as_view(), name="token_verify"),
]
urlpatterns = [
    path("jwt/", include(jwt_urlpatterns)),
    path("user-details/", UserDetailAPIView.as_view(), name="user-details"),
]
