requirements/base.txt
crispy-bootstrap5~=0.6
Django~=4.1.1
django-allauth~=0.51.0
django-crispy-forms~=1.14.0
django-environ~=0.9.0
django-extensions~=3.2.1
djangorestframework~=3.14.0
djangorestframework-simplejwt~=5.2.2
psycopg2-binary~=2.9.3
whitenoise~=6.2.0
